Added structure event detail screen

+++ app/src/main/java/br/com/teste/sicredi/feature/events/ui/EventsAdapter.kt
@@ -11,7 +11,7 @@ import kotlinx.android.synthetic.main.item_events.view.*
 
 class EventsAdapter(
     private val items: List<EventsData>,
-    private val listener: (Double) -> Unit
+    private val listener: (Int) -> Unit
 ) : RecyclerView.Adapter<EventsAdapter.ViewHolder>() {
 
     inner class ViewHolder(parent: ViewGroup) :
